# docs/swagger.yaml
basePath: /api
definitions:
  api.SignUpRequest:
    properties:
      avatarURL:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      lastName:
        type: string
      location:
        type: string
      organizationID:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      provider:
        type: string
      providerID:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  api.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      gender:
        type: string
      is_active:
        type: boolean
      is_superuser:
        type: boolean
      last_name:
        type: string
      locale:
        type: string
      phone_number:
        type: string
      phone_verified:
        type: boolean
      username:
        type: string
    type: object
  gin.H:
    additionalProperties: {}
    type: object
  github_com_malekradhouane_trippy_pkg_interfaces.Credential:
    properties:
      password:
        example: "1234"
        type: string
      username:
        example: "042"
        type: string
    type: object
  github_com_malekradhouane_trippy_pkg_interfaces.Token:
    properties:
      accessToken:
        description: token
        type: string
      expiration:
        description: expiration date
        example: ""
        type: string
      refreshToken:
        description: refresh token to be used for renewal
        type: string
    type: object
  github_com_malekradhouane_trippy_pkg_interfaces.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      date_of_birth:
        description: Using string pointer for optional date
        type: string
      email:
        description: Authentication
        type: string
      email_verified:
        description: Account status
        type: boolean
      first_name:
        type: string
      gender:
        type: string
      id:
        description: Primary key
        type: string
      is_active:
        type: boolean
      is_superuser:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      last_password_change:
        type: string
      locale:
        type: string
      metadata:
        additionalProperties: true
        description: Metadata
        type: object
      mfa_enabled:
        description: Security
        type: boolean
      phone_number:
        type: string
      phone_verified:
        type: boolean
      provider:
        type: string
      provider_id:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        description: User profile
        type: string
    type: object
  github_com_malekradhouane_trippy_utils_httpresp.HTTPError:
    properties:
      code:
        description: the HTTP status
        example: 500
        type: integer
      error:
        description: the error description
        example: internal server error
        type: string
      success:
        description: wether the business action is successfull or not
        example: false
        type: boolean
    type: object
  github_com_malekradhouane_trippy_utils_httpresp.HTTPError401:
    properties:
      code:
        description: the HTTP status
        example: 401
        type: integer
      error:
        description: the error description
        example: unauthorized
        type: string
      success:
        description: wether the business action is successfull or not
        example: false
        type: boolean
    type: object
  github_com_malekradhouane_trippy_utils_httpresp.JSONResult:
    properties:
      code:
        description: the HTTP status
        example: 200
        type: integer
      result:
        description: whatever object or array of objects
      success:
        description: wether the business action is successfull or not
        example: true
        type: boolean
    type: object
host: localhost:5001
info:
  contact:
    email: malek.radhouen@gmail.com
    name: API Support
    url: https://www.trippy.fr/support
  description: Here is our solution documentation and testing portal of provided functionalities
    to interact with our hypervisor tool.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.trippy.com/terms/
  title: Trippy - Inventory solution API
  version: "1.0"
paths:
  /activate/{token}:
    get:
      consumes:
      - application/json
      description: Activates a user account using the validation token from email
      operationId: activate-account
      parameters:
      - description: Activation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account activated successfully
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.JSONResult'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Activate user account
      tags:
      - authenticate
  /authenticate:
    post:
      consumes:
      - application/json
      description: authenticate user using provided credentials. Returned token can
        be used as "Bearer [insert token here]" in the 'Authorize' form
      operationId: authenticate-user-by-username-password
      parameters:
      - description: Authentication credentials
        in: body
        name: credential
        required: true
        schema:
          $ref: '#/definitions/github_com_malekradhouane_trippy_pkg_interfaces.Credential'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.JSONResult'
            - properties:
                result:
                  $ref: '#/definitions/github_com_malekradhouane_trippy_pkg_interfaces.Token'
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Authenticate user
      tags:
      - authenticate
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset email to the user
      operationId: forgot-password
      parameters:
      - description: Email address
        in: body
        name: email
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.JSONResult'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Request password reset
      tags:
      - authenticate
  /logout:
    post:
      consumes:
      - application/json
      description: Logs out the authenticated user by invalidating the token
      operationId: logout-user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.JSONResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError401'
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - authenticate
  /refresh_token:
    post:
      consumes:
      - application/json
      description: refreshes the provided token
      operationId: refresh-token
      parameters:
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.JSONResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: RefreshToken refreshes authenticated user's token
      tags:
      - authenticate
  /reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using the token
      operationId: reset-password
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.JSONResult'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Reset password
      tags:
      - authenticate
  /status:
    get:
      description: Health check endpoint returning service status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Service status
      tags:
      - status
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/github_com_malekradhouane_trippy_pkg_interfaces.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_pkg_interfaces.User'
        "400":
          description: Data validation error
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_pkg_interfaces.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Get a user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_pkg_interfaces.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_malekradhouane_trippy_utils_httpresp.HTTPError'
      summary: Update a user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
x-extension-openapi:
  example: value on a json format